Refactor: Novas melhorias  - Refatoração de comportamentos para melhoria do entendimento.
Marcos Garcia
