+++ desktop/src/main/java/com/github/macgarcia/access/control/desktop/component/ModeloIntegracao.java
@@ -3,7 +3,7 @@ package com.github.macgarcia.access.control.desktop.component;
 import com.github.macgarcia.access.control.desktop.configuration.Configuracao;
 import com.github.macgarcia.access.control.desktop.configuration.FactoryLog;
 import com.github.macgarcia.access.control.desktop.integracao.ClientApiBackup;
-import com.github.macgarcia.access.control.desktop.integracao.ConfiguracaoIntegracao;
+import com.github.macgarcia.access.control.desktop.model.ConfiguracaoIntegracao;
 import com.github.macgarcia.access.control.desktop.integracao.HistoricoDtoRequest;
 import com.github.macgarcia.access.control.desktop.integracao.IntegracaoResponse;
 import com.github.macgarcia.access.control.desktop.integracao.NotaDtoRequest;
@@ -12,7 +12,6 @@ import com.github.macgarcia.access.control.desktop.model.Nota;
 import com.github.macgarcia.access.control.desktop.repository.ConfiguracaoIntegracaoRepository;
 import com.github.macgarcia.access.control.desktop.repository.NotaRepository;
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
