+++ desktop/src/main/java/com/github/macgarcia/access/control/desktop/view/configuracoes/TelaConfigurarIntegracaoApi.java
@@ -2,7 +2,7 @@ package com.github.macgarcia.access.control.desktop.view.configuracoes;
 
 import com.github.macgarcia.access.control.desktop.component.ModeloIntegracao;
 import com.github.macgarcia.access.control.desktop.configuration.FactoryMensagem;
-import com.github.macgarcia.access.control.desktop.integracao.ConfiguracaoIntegracao;
+import com.github.macgarcia.access.control.desktop.model.ConfiguracaoIntegracao;
 import com.github.macgarcia.access.control.desktop.enuns.FlagIntegracao;
 import com.github.macgarcia.access.control.desktop.repository.ConfiguracaoIntegracaoRepository;
 import java.util.Objects;
